feat: Add useUserAuth hook for user authentication
This commit adds a new file, userAuth.js, which contains the useUserAuth hook. This hook provides a loginUser function that sends a POST request to the /api/v1/users/auth endpoint with the provided email and password. If the response is successful, the received token and current user ID are stored in the local storage. Otherwise, an error is thrown with the error message from the response.