refactor: Update custom.css to add flex container and right panel styles
This commit updates the custom.css file to add styles for a flex container and a right panel. The flex container is set to display as a flexbox and have a height of 100vh. The right panel container is given a flex value of 1 and padding of 20px. These changes will be used to create a layout for the application.