feat: Add useDepartmentAdd hook for adding departments
This commit adds a new hook called useDepartmentAdd, which provides a function for adding departments. The hook makes a POST request to the "/api/v1/departments" endpoint with the provided form data and the user's authorization token. If the request is successful, the response data is returned. If the request fails, an error is thrown with the error message from the response.