feat: Add checkAuth function to verify user authentication
This commit adds a new file, checkAuth.js, which contains the checkAuth function. This function checks if the user is logged in by checking for the presence of a token in the local storage. If the user is not logged in, it returns an object with a status code of 401 and a message indicating that the user must be logged in to view the page. Otherwise, it returns null, indicating that the user is authenticated.